# docs/vault_swagger.yaml
definitions:
  services.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  services.Resource:
    properties:
      created_at:
        type: string
      error:
        type: string
      resource_id:
        type: string
      status:
        $ref: '#/definitions/services.Status'
      stored_size:
        type: integer
      total_size:
        type: integer
      updated_at:
        type: string
    type: object
  services.Status:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - StatusQueuedForStoring
    - StatusStoring
    - StatusStored
    - StatusStoreError
    - StatusQueuedForDeletion
    - StatusDeleting
    - StatusDeleteError
info:
  contact:
    email: support@webtor.io
    name: Webtor Support
    url: https://webtor.io/support
  description: API to communicate with Vault service.
  title: Vault API
  version: "0.1"
paths:
  /resource/{id}:
    delete:
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/services.Resource'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Queue deletion of a resource
      tags:
      - resource
    get:
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Resource'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Get resource
      tags:
      - resource
    put:
      description: Creates the resource if missing or marks it queued for processing
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/services.Resource'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Queue storing of a resource
      tags:
      - resource
  /webseed/{id}/{path}:
    get:
      description: Proxies stored files from S3 with Range support. Returns 404 if
        resource is not fully stored or file not found.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Path inside resource
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "206":
          description: Partial Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Webseed proxy
      tags:
      - webseed
    head:
      description: Proxies stored files from S3 with Range support. Returns 404 if
        resource is not fully stored or file not found.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Path inside resource
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "206":
          description: Partial Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Webseed proxy
      tags:
      - webseed
swagger: "2.0"